consolidate autoLauncher class and os functions into one file outside of native.ts
