+++ src/lib/global.ts
@@ -1,7 +1,7 @@
 import { reactive } from "vue";
-import {getLang,LangType} from "../loc/lang"
+import { getLang, LangType } from "../loc/lang"
 import { Client } from "src/lib/client"
-import { AutoLauncher } from "src/lib/native"
+import { AutoLauncher } from "src/lib/autoLauncher"
 
 const text: LangType = {}
 // TODO use dependency injection to ensure methods and properties can't be accessed unless they are initialized and valid
@@ -9,14 +9,14 @@ export class Global {
   client = new Client
   autoLauncher = new AutoLauncher
   data = reactive({ status: { state: "loading", message: "loading" }, loc: { selected: 'en', text } })
-  async changeLang(newLang: string):Promise<void> {
+  async changeLang(newLang: string): Promise<void> {
     this.data.loc.selected = newLang
     await this.loadLangData()
   }
-  async loadLangData():Promise<void> {
+  async loadLangData(): Promise<void> {
     this.data.loc.text = await getLang(this.data.loc.selected)
   }
-  async init():Promise<void> {
+  async init(): Promise<void> {
     await Promise.all(
       [
         this.autoLauncher.init(),
